+++ src/camera_util.c
@@ -1,4 +1,4 @@
-#include "camera_transform.h"
+#include "camera_util.h"
 
 // Returns the screen length of a distance `length` at distance `z` with respect to the camera.
 int GetRenderLength(GameData* GD, int length, fixed_t z) {
@@ -11,8 +11,8 @@ void GetRenderCoords(GameData* GD, fixed_t x, fixed_t y, fixed_t z, int* ox, int
   fixed_t ty = y - GD->camera.y;
 
   // zoom
-  int itx = (tx * GD->camera.zoom / fixed_factor / z);
-  int ity = (ty * GD->camera.zoom / fixed_factor / z);
+  int itx = GetRenderLength(GD, tx, z) / fixed_factor;
+  int ity = GetRenderLength(GD, ty, z) / fixed_factor;
 
   *ox = itx + render_w / 2;
   *oy = ity + render_h / 2;
