+++ src/camera_util.h
