+++ src/shape_util.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "gd.h"
+
+void SpawnNewShapes(GameData* GD);
+int CompareShapes(const void* p, const void* q);
+int ClaimEmptyShapeSlot(GameData* GD);
+int PickShapeSides(GameData* GD);
+ShapeVariant PickShapeVariant(GameData* GD, int sides);
+void SpawnChildShapes(GameData* GD, int parent);
